broke out circuitboard, but http error 422 sendcircuit prob wrong deserialize json
